0.17 cherry-picks

+++ tests/integration/modules/file.py
@@ -42,7 +42,8 @@ class FileModuleTest(integration.ModuleCase):
         super(FileModuleTest, self).setUp()
 
     def tearDown(self):
-        os.remove(self.myfile)
+        if os.path.isfile(self.myfile):
+            os.remove(self.myfile)
         if os.path.islink(self.mysymlink):
             os.remove(self.mysymlink)
         if os.path.islink(self.mybadsymlink):
@@ -154,7 +155,7 @@ class FileModuleTest(integration.ModuleCase):
         self.assertTrue(ret)
 
     def test_cannot_remove(self):
-        ret = self.run_function('file.remove', arg=['/dev/tty'])
+        ret = self.run_function('file.remove', arg=['tty'])
         self.assertEqual(
             'ERROR executing file.remove: File path must be absolute.', ret
         )
